+++ src/polyzymd/analyses/mda/__init__.py
@@ -5,16 +5,6 @@ that wrap MDAnalysis ``AnalysisBase`` objects without importing MDAnalysis at
 module import time.
 """
 
-from polyzymd.analyses.mda.aggregation import (
-    AggregatedMetric,
-    ExplicitReplicateMetricPolicy,
-    MDAAggregationContext,
-    MDAAggregationError,
-    ReplicateMetricPolicy,
-    aggregate_replicate_artifacts,
-    aggregate_replicate_artifacts_from_disk,
-    validate_autocorrelation_estimator_version,
-)
 from polyzymd.analyses.mda.artifacts import (
     MDA_ARTIFACT_SCHEMA_VERSION,
     ArtifactEnvelope,
@@ -39,15 +29,8 @@ from polyzymd.analyses.mda.job import (
     MDAJobResult,
     MDAUniversePolicy,
 )
-from polyzymd.analyses.mda.lifecycle import MDAReplicateJobContext
+from polyzymd.analyses.mda.lifecycle import MDAReplicateJobContext, frame_selection_payload
 from polyzymd.analyses.mda.pair_distance import PairSelection, pair_distance_matrix
-from polyzymd.analyses.mda.plugin import (
-    MDAArtifactCollector,
-    MDACollectorContext,
-    StrictJSONMDAResultCollector,
-    frame_selection_payload,
-    strict_json_payload,
-)
 from polyzymd.analyses.mda.store import ArtifactStore, ArtifactStoreError
 from polyzymd.analyses.mda.universe import FileIdentity, UniverseProvenance, UniverseProvider
 
@@ -57,14 +40,6 @@ __all__ = [
     "AnalysisBaseLike",
     "MDAnalysisExtensionError",
     "MDARunKwargs",
-    "AggregatedMetric",
-    "ExplicitReplicateMetricPolicy",
-    "MDAAggregationContext",
-    "MDAAggregationError",
-    "validate_autocorrelation_estimator_version",
-    "ReplicateMetricPolicy",
-    "aggregate_replicate_artifacts",
-    "aggregate_replicate_artifacts_from_disk",
     "ArtifactEnvelope",
     "ArtifactManifest",
     "ArtifactSidecarRef",
@@ -83,11 +58,7 @@ __all__ = [
     "MDAReplicateJobContext",
     "PairSelection",
     "pair_distance_matrix",
-    "MDAArtifactCollector",
-    "MDACollectorContext",
-    "StrictJSONMDAResultCollector",
     "frame_selection_payload",
-    "strict_json_payload",
     "FileIdentity",
     "UniverseProvider",
     "UniverseProvenance",